Move get_traded_orders into DriverLogger
